Add files via upload
col1=district number. the district number is a hyperlink to each district's Wikipedia site.
col2 =location. the location is a map with a hyperlink to the Wikipedia district map.
I think it might be wise to move location to elected_officials since that is the location reference. We are not referencing the location of the donors.