Adding file with crawler for vote gathering
Not complete yet, still working on figuring out how to choose only the most recent date. Unfortunately the data is not consistently in the same order (as far as chamber goes) and sometimes one chamber has multiple votes on a bill but the other does not. So, it is presenting some difficulty. This is what I have so far.