Completed crawler w/ csv output file
Ok, so I completed my crawler and got votes for a little under 3000 bills. I pulled the bill numbers from the csv file in raw data called "billsbysponsor.csv." Also, included is the csv generated from my crawler, with a header included. I can remove the header if its not needed and/or will get in the way of converting it to a table.