Donor Crawler Update
Donor crawler has been updated to gather a guessed role using a heuristics algorithm in addition to the name of the donor. No other fields are considered anymore. Donors are now marked as politicians if the algorithm determines that.
Layout has been tweaked slightly since this afternoon to reflect that change