Add files via upload
This does not have the technical summary because I would like to leave the technical parts to the showcase. Way too much code to write a half decent technical write up on, so I felt it would be best to pick the pieces we want and go from there. Let me know if you guys see and errors here and if you have changes you want made to the layout. Still editing.